+++ crates/biome_js_analyze/tests/specs/nursery/noInteractiveElementToNoninteractiveRole/invalid.jsx
@@ -1,0 +1,66 @@
+/* Interactive elements */
+<a href="http://x.y.z" role="img" />;
+<a href="http://x.y.z" tabIndex="0" role="img" />;
+/* All flavors of input */
+<input role="img" />;
+<input type="img" role="img" />;
+<input type="checkbox" role="img" />;
+<input type="color" role="img" />;
+<input type="date" role="img" />;
+<input type="datetime" role="img" />;
+<input type="datetime-local" role="img" />;
+<input type="email" role="img" />;
+<input type="file" role="img" />;
+<input type="image" role="img" />;
+<input type="month" role="img" />;
+<input type="number" role="img" />;
+<input type="password" role="img" />;
+<input type="radio" role="img" />;
+<input type="range" role="img" />;
+<input type="reset" role="img" />;
+<input type="search" role="img" />;
+<input type="submit" role="img" />;
+<input type="tel" role="img" />;
+<input type="text" role="img" />;
+<input type="time" role="img" />;
+<input type="url" role="img" />;
+<input type="week" role="img" />;
+/* End all flavors of input */
+<menuitem role="img" />;
+<option className="foo" role="img" />;
+<select className="foo" role="img" />;
+<textarea className="foo" role="img" />;
+<tr role="img" />;
+/* Interactive elements */
+<a href="http://x.y.z" role="listitem" />;
+<a href="http://x.y.z" tabIndex="0" role="listitem" />;
+/* All flavors of input */
+<input role="listitem" />;
+<input type="listitem" role="listitem" />;
+<input type="checkbox" role="listitem" />;
+<input type="color" role="listitem" />;
+<input type="date" role="listitem" />;
+<input type="datetime" role="listitem" />;
+<input type="datetime-local" role="listitem" />;
+<input type="email" role="listitem" />;
+<input type="file" role="listitem" />;
+<input type="image" role="listitem" />;
+<input type="month" role="listitem" />;
+<input type="number" role="listitem" />;
+<input type="password" role="listitem" />;
+<input type="radio" role="listitem" />;
+<input type="range" role="listitem" />;
+<input type="reset" role="listitem" />;
+<input type="search" role="listitem" />;
+<input type="submit" role="listitem" />;
+<input type="tel" role="listitem" />;
+<input type="text" role="listitem" />;
+<input type="time" role="listitem" />;
+<input type="url" role="listitem" />;
+<input type="week" role="listitem" />;
+/* End all flavors of input */
+<menuitem role="listitem" />;
+<option className="foo" role="listitem" />;
+<select className="foo" role="listitem" />;
+<textarea className="foo" role="listitem" />;
+<tr role="listitem" />;